requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.10.7
django-filter==1.0.4
django-haystack==2.6.1
djangorestframework==3.6.3
elasticsearch==2.4.1
idna==2.5
Markdown==2.6.8
pyelasticsearch==1.4
python-memcached==1.58
pytz==2017.2
simplejson==3.11.1
six==1.10.0
urllib3==1.21.1
